requirement.txt
alabaster==0.7.9
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.23
coverage==4.2
coveralls==1.1
decorator==4.0.10
Django==1.9.7
djangorestframework==3.4.4
docopt==0.6.2
docutils==0.12
fancycompleter==0.5
funcsigs==1.0.2
future==0.15.2
gunicorn==19.6.0
imagesize==0.7.1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
mock==2.0.0
ordereddict==1.1
pathlib2==2.1.0
pbr==1.10.0
pdbpp==0.8.3
pexpect==4.2.0
pickleshare==0.7.4
pluggy==0.3.1
prompt-toolkit==1.0.6
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
Pygments==2.1.3
pyrepl==0.8.4
pytest==2.9.2
pytest-cov==2.3.1
pytest-mock==1.2
pytest-pythonpath==0.7
pytz==2016.6.1
requests==2.11.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.5
sphinx-rtd-theme==0.1.9
testfixtures==4.10.0
tox==2.3.1
traitlets==4.2.2
virtualenv==15.0.3
wcwidth==0.1.7
websocket-client==0.37.0
wmctrl==0.3
